node query function to utils
